+++ mi/instrument/kut/ek60/ooicore/zplsc_echogram.py
@@ -6,18 +6,14 @@
 
 Release notes:
 
-This class supports the generation of ZPLSC echograms. It needs matplotlib version 1.3.1 for the code to display the
-colorbar at the bottom of the figure. If matplotlib version 1.1.1 is used, the colorbar is plotted over the
-figure instead of at the bottom of it.
+This class supports the generation of ZPLSC echograms.
 """
 
 import matplotlib
 
 matplotlib.use("Agg")
 
-import matplotlib.pyplot as plt
-from matplotlib.dates import date2num, num2date
-from modest_image import imshow
+from matplotlib.dates import date2num
 
 from datetime import datetime
 
@@ -101,7 +97,6 @@ def read_config_header(chunk):
     values.pop(4)  # drop the spare field
 
     # strip the trailing zero byte padding from the strings
-    # for i in [0, 1, 2, 3]:
     for i in xrange(4):
         values[i] = values[i].strip('\x00')
 
@@ -146,129 +141,3 @@ def read_config_transducer(chunk):
     config_transducer[field_names[19]] = sa_correction_table
     config_transducer[field_names[20]] = values[35]
     return config_transducer
-
-
-class ZPLSPlot(object):
-    font_size_small = 14
-    font_size_large = 18
-    num_xticks = 25
-    num_yticks = 7
-    interplot_spacing = 0.1
-    lower_percentile = 5
-    upper_percentile = 95
-
-    def __init__(self, data_times, power_data_dict, frequency_dict, bin_size):
-        self.power_data_dict = self._transpose_and_flip(power_data_dict)
-        self.min_db, self.max_db = self._get_power_range(power_data_dict)
-        self.frequency_dict = frequency_dict
-
-        # convert ntp time, i.e. seconds since 1900-01-01 00:00:00 to matplotlib time
-        self.data_times = (data_times / (60 * 60 * 24)) + REF_TIME
-        max_depth, _ = self.power_data_dict[1].shape
-        self._setup_plot(bin_size, max_depth)
-
-    def generate_plots(self):
-        """
-        Generate plots for all transducers in data set
-        """
-        freq_to_channel = {v: k for k, v in self.frequency_dict.iteritems()}
-        data_axes = None
-        for index, frequency in enumerate(sorted(freq_to_channel)):
-            channel = freq_to_channel[frequency]
-            td_f = self.frequency_dict[channel]
-            title = 'Power: Transducer #%d: Frequency: %0.1f kHz' % (channel, td_f / 1000)
-            data_axes = self._generate_plot(self.ax[index], self.power_data_dict[channel], title,
-                                            self.min_db, self.max_db)
-
-        if data_axes:
-            self._display_x_labels(self.ax[2], self.data_times)
-            self.fig.tight_layout(rect=[0, 0.0, 0.97, 1.0])
-            self._display_colorbar(self.fig, data_axes)
-
-    def write_image(self, filename):
-        self.fig.savefig(filename)
-        plt.close(self.fig)
-        self.fig = None
-
-    def _setup_plot(self, bin_size, max_depth):
-        # subset the yticks so that we don't plot every one
-        yticks = np.linspace(0, max_depth, self.num_yticks)
-        # create range vector (depth in meters)
-        yticklabels = np.round(np.linspace(0, max_depth * bin_size, self.num_yticks)).astype(int)
-
-        self.fig, self.ax = plt.subplots(len(self.frequency_dict), sharex=True, sharey=True)
-        self.fig.subplots_adjust(hspace=self.interplot_spacing)
-        self.fig.set_size_inches(40, 19)
-
-        for axes in self.ax:
-            axes.grid(False)
-            axes.set_ylabel('depth (m)', fontsize=self.font_size_small)
-            axes.set_yticks(yticks)
-            axes.set_yticklabels(yticklabels, fontsize=self.font_size_small)
-            axes.tick_params(axis="both", labelcolor="k", pad=4, direction='out', length=5, width=2)
-            axes.spines['top'].set_visible(False)
-            axes.spines['right'].set_visible(False)
-            axes.spines['bottom'].set_visible(False)
-            axes.spines['left'].set_visible(False)
-
-    @staticmethod
-    def _get_power_range(power_dict):
-        # Calculate the power data range across all channels
-        all_power_data = np.concatenate(power_dict.values())
-        max_db = np.nanpercentile(all_power_data, ZPLSPlot.upper_percentile)
-        min_db = np.nanpercentile(all_power_data, ZPLSPlot.lower_percentile)
-        return min_db, max_db
-
-    @staticmethod
-    def _transpose_and_flip(power_dict):
-        for channel in power_dict:
-            # Transpose array data so we have time on the x-axis and depth on the y-axis
-            power_dict[channel] = power_dict[channel].transpose()
-            # reverse the Y axis (so depth is measured from the surface (at the top) to the ZPLS (at the bottom)
-            power_dict[channel] = power_dict[channel][::-1]
-        return power_dict
-
-    @staticmethod
-    def _generate_plot(ax, power_data, title, min_db, max_db):
-        """
-        Generate a ZPLS plot for an individual channel
-        :param ax:  matplotlib axis to receive the plot image
-        :param power_data:  Transducer data array
-        :param data_times:  Transducer internal time array
-        :param title:  plot title
-        :param min_db: minimum power level
-        :param max_db: maximum power level
-        """
-        # only generate plots for the transducers that have data
-        if power_data.size <= 0:
-            return
-
-        ax.set_title(title, fontsize=ZPLSPlot.font_size_large)
-        return imshow(ax, power_data, interpolation='none', aspect='auto', cmap='jet', vmin=min_db, vmax=max_db)
-
-    @staticmethod
-    def _display_x_labels(ax, data_times):
-        time_format = '%Y-%m-%d\n%H:%M:%S'
-        time_length = data_times.size
-        # X axis label
-        # subset the xticks so that we don't plot every one
-        xticks = np.linspace(0, time_length, ZPLSPlot.num_xticks)
-        xstep = int(round(xticks[1]))
-        # format trans_array_time array so that it can be used to label the x-axis
-        xticklabels = [i for i in num2date(data_times[::xstep])] + [num2date(data_times[-1])]
-        xticklabels = [i.strftime(time_format) for i in xticklabels]
-
-        # rotates and right aligns the x labels, and moves the bottom of the
-        # axes up to make room for them
-        ax.set_xlabel('time (UTC)', fontsize=ZPLSPlot.font_size_small)
-        ax.set_xticks(xticks)
-        ax.set_xticklabels(xticklabels, rotation=45, horizontalalignment='center', fontsize=ZPLSPlot.font_size_small)
-        ax.set_xlim(0, time_length)
-
-    @staticmethod
-    def _display_colorbar(fig, data_axes):
-        # Add a colorbar to the specified figure using the data from the given axes
-        ax = fig.add_axes([0.965, 0.12, 0.01, 0.775])
-        cb = fig.colorbar(data_axes, cax=ax, use_gridspec=True)
-        cb.set_label('dB', fontsize=ZPLSPlot.font_size_large)
-        cb.ax.tick_params(labelsize=ZPLSPlot.font_size_small)